Move test_compare_json to its own file
Avoid it getting collected multiple times by every file that imports the
helpers module.
