+++ server/tests/helpers.py
@@ -1,7 +1,6 @@
 import uuid, json, re
 from datetime import datetime
 from typing import Any, List, Union, Tuple
-import pytest
 from flask.testing import FlaskClient
 from werkzeug.wrappers import Response
 from sqlalchemy.orm import joinedload
@@ -272,50 +271,3 @@ def compare_json(actual_json, expected_json):
             ), f"Actual: {actual_json}\nExpected: {expected_json}\nKeypath: {serialize_keypath(current_keypath)}"
 
     inner_compare_json(actual_json, expected_json, [])
-
-
-def _test_compare_json():
-    def asserts_gt(num: int):
-        def assert_gt(value: int):
-            assert isinstance(value, int)
-            assert value > num
-
-        return assert_gt
-
-    compare_json([], [])
-    compare_json({}, {})
-    compare_json("a", "a")
-    compare_json(1, 1)
-    compare_json([1, {}], [1, {}])
-    compare_json(1, asserts_gt(0))
-
-    with pytest.raises(AssertionError, match=r"Actual: 1\nExpected: 2\nKeypath: root"):
-        compare_json(1, 2)
-
-    with pytest.raises(
-        AssertionError, match=r'Actual: 1\nExpected: 2\nKeypath: root\["a"\]'
-    ):
-        compare_json({"a": 1}, {"a": 2})
-
-    with pytest.raises(AssertionError, match=r"dict keys do not match at root"):
-        compare_json({"a": 1}, {})
-
-    with pytest.raises(AssertionError, match=r"expected dict, got list at root"):
-        compare_json([], {})
-
-    with pytest.raises(AssertionError, match=r"expected list, got dict at root"):
-        compare_json({}, [])
-
-    with pytest.raises(AssertionError, match=r"list lengths do not match at root"):
-        compare_json([1], [])
-
-    with pytest.raises(
-        AssertionError, match=r"Actual: 2\nExpected: 3\nKeypath: root\[1\]"
-    ):
-        compare_json([1, 2], [1, 3])
-
-    with pytest.raises(AssertionError, match=r"custom comparison failed at root"):
-        compare_json(1, asserts_gt(1))
-
-    with pytest.raises(AssertionError, match=r"custom comparison failed at root\[2\]"):
-        compare_json([1, 2, 3], [asserts_gt(0), asserts_gt(1), asserts_gt(3)])
